requirements.txt
conda-repo-cli==1.0.75
conda-verify==3.4.2
pandas-dq==1.29
scikit-learn==1.3.0
scikit-learn-intelex==20230426.121932
scipy==1.11.1
streamlit==1.31.0

